Poll for changes independently from websockets
If `--action`s are passed to the app, a worker will be started in
background in order to poll for changes and execute those actions, so that
this happens independently from the notifications service.

Decreases polling interval from 500ms to 200ms.

Fixes #9. Actions will also get executed when there are no active
websocket connections.